% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_infer.R
\name{haplotype_infer}
\alias{haplotype_infer}
\title{Infer parental transmitted and non-transmitted alleles}
\usage{
haplotype_infer(vcf_dt, hap_length = 5e+05)
}
\arguments{
\item{vcf_dt}{A `data.table` containing fixed VCF columns and genotype
columns named `M`, `P`, and `B`.}

\item{hap_length}{Integer, haplotype window length.}
}
\value{
A named list containing:
\describe{
  \item{vcf_trans}{A `data.table` with transmitted alleles.}
  \item{vcf_nontrans}{A `data.table` with non-transmitted alleles.}
  \item{sim_perc_summary}{A `data.table` summarizing inference statistics.}
}
}
\description{
Given trio genotype data for a single family, infer maternal and paternal
transmitted and non-transmitted alleles.
}
\examples{
vcf_file <- system.file("extdata", "Toy_TrioGenotype.vcf.gz", package = "parati")
vcf_dt <- read_vcf_by_chr(vcf_file, chr = 1)
vcf_sub <- vcf_dt[, c(names(vcf_dt)[1:9], "1-M", "1-P", "1-B"), with = FALSE]
data.table::setnames(vcf_sub, c("1-M", "1-P", "1-B"), c("M", "P", "B"))
res <- haplotype_infer(vcf_sub)
names(res)

}
