% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{read_vcf_by_chr}
\alias{read_vcf_by_chr}
\title{Read VCF by chromosome}
\usage{
read_vcf_by_chr(vcf_file, chr)
}
\arguments{
\item{vcf_file}{Character scalar, path to a VCF/VCF.GZ file.}

\item{chr}{Character or integer chromosome identifier.}
}
\value{
A `data.table` containing VCF rows for the selected chromosome.
}
\description{
Reads a VCF file and subsets variants to a given chromosome.
}
\examples{
vcf_file <- system.file("extdata", "Toy_TrioGenotype.vcf.gz", package = "parati")
vcf_chr <- read_vcf_by_chr(vcf_file, chr = 1)
dim(vcf_chr)

}
