% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_plink.R
\name{vcf_to_plink}
\alias{vcf_to_plink}
\title{Export VCF to PLINK}
\usage{
vcf_to_plink(vcf_trans, vcf_nontrans, out_dir, chr, plink_path)
}
\arguments{
\item{vcf_trans}{vcfR object of transmitted alleles}

\item{vcf_nontrans}{vcfR object of non-transmitted alleles}

\item{out_dir}{character, output directory path}

\item{chr}{chromosome identifier}

\item{plink_path}{character, path to PLINK executable}
}
\value{
Invisibly returns \code{NULL}. The main effect of the function is writing PLINK files to the specified `out_dir`.
}
\description{
Export VCF to PLINK
}
\details{
Converts transmitted and non-transmitted VCF objects into PLINK format files.
}
