% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{write_vcf_dt}
\alias{write_vcf_dt}
\title{Write VCF data.table to file.}
\usage{
write_vcf_dt(df, file)
}
\arguments{
\item{df}{data.table containing VCF rows}

\item{file}{output file path}
}
\value{
Invisibly returns \code{NULL}. The VCF file is written to \code{file}.
}
\description{
Write a data.table representing VCF rows into a VCF file.
}
