% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{write_vcf_obj}
\alias{write_vcf_obj}
\title{Write vcfR object to file}
\usage{
write_vcf_obj(vcf_obj, file)
}
\arguments{
\item{vcf_obj}{A `vcfR` object.}

\item{file}{Character scalar, output file path.}
}
\value{
`NULL`, invisibly. The VCF file is written to `file`.
}
\description{
Writes a `vcfR` object to a VCF file.
}
\examples{
vcf_file <- system.file("extdata", "Toy_TrioGenotype.vcf.gz", package = "parati")
vcf_dt <- read_vcf_by_chr(vcf_file, chr = 1)
vcf_obj <- vcf_dt_to_vcfR(vcf_dt)
outfile <- tempfile(fileext = ".vcf")
write_vcf_obj(vcf_obj, outfile)
file.exists(outfile)

}
