% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{read_vcf_by_chr}
\alias{read_vcf_by_chr}
\title{Read VCF by chromosome}
\usage{
read_vcf_by_chr(vcf_file, chr)
}
\arguments{
\item{vcf_file}{character VCF file path}

\item{chr}{character/integer Chromosome}
}
\value{
data.table
}
\description{
Read VCF by chromosome
}
