% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{vcf_dt_to_vcfR}
\alias{vcf_dt_to_vcfR}
\title{Convert data.table to vcfR object}
\usage{
vcf_dt_to_vcfR(df, meta = character())
}
\arguments{
\item{df}{A `data.table` containing VCF rows.}

\item{meta}{Character vector of VCF meta lines.}
}
\value{
A `vcfR` object.
}
\description{
Converts a `data.table` representing VCF rows into a `vcfR` object.
}
\examples{
vcf_file <- system.file("extdata", "Toy_TrioGenotype.vcf.gz", package = "parati")
vcf_dt <- read_vcf_by_chr(vcf_file, chr = 1)
vcf_obj <- vcf_dt_to_vcfR(vcf_dt)
class(vcf_obj)
stopifnot(inherits(vcf_obj, "vcfR"))

}
