% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{vcf_dt_to_vcfR}
\alias{vcf_dt_to_vcfR}
\title{Convert data.table to vcfR object}
\usage{
vcf_dt_to_vcfR(df, meta)
}
\arguments{
\item{df}{data.table containing VCF rows}

\item{meta}{list of meta information for VCF}
}
\value{
vcfR object
}
\description{
Converts a data.table representing VCF rows into a vcfR object.
}
