% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{write_vcf_dt}
\alias{write_vcf_dt}
\title{Convert VCF data.table to vcfR object}
\usage{
write_vcf_dt(df, file)
}
\arguments{
\item{df}{data.table containing VCF rows}

\item{file}{output file path}
}
\description{
Converts a data.table representing VCF rows into a vcfR object.
}
