% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_vcf.R
\name{write_vcf_obj}
\alias{write_vcf_obj}
\title{Write vcfR object to file}
\usage{
write_vcf_obj(vcf_obj, file)
}
\arguments{
\item{vcf_obj}{vcfR object}

\item{file}{output file path}
}
\description{
Writes a vcfR object to a VCF file.
}
